Moved legacy_msg_reflection.{c,h} -> reflection.{c.h}.
